kernel: [NETLINK]: Fix trivial signed/unsigned compiler warnings
Also, kernel.org commits
 838965ba22066c7fcdbacfc543c387d0eb76c14c [NETLINK]: Add NLA_PUT_BE16/nla_get_be16()
 1045b03e07d85f3545118510a587035536030c1c netlink: fix overrun in attribute iteration
 619e803d3c1b7bcc17c45e81f309d0b9b3df2d5d netlink: fix (theoretical) overrun in message iteration

+++ linux/linux-2.6.22.19/include/linux/netdevice.h
@@ -862,10 +862,10 @@ enum {
 #define netif_msg_hw(p)		((p)->msg_enable & NETIF_MSG_HW)
 #define netif_msg_wol(p)	((p)->msg_enable & NETIF_MSG_WOL)
 
-static inline u32 netif_msg_init(int debug_value, int default_msg_enable_bits)
+static inline u32 netif_msg_init(int debug_value, u32 default_msg_enable_bits)
 {
 	/* use default */
-	if (debug_value < 0 || debug_value >= (sizeof(u32) * 8))
+	if (debug_value < 0 || debug_value >= (int)(sizeof(u32) * 8))
 		return default_msg_enable_bits;
 	if (debug_value == 0)	/* no output */
 		return 0;
